add .gitignore gile to exclude settings

+++ source/ztf_image_pipeline.py
@@ -5,6 +5,7 @@ Build a .JSON meta file for each object.
 '''
 
 import os
+import sys
 import re
 import subprocess
 import errno
@@ -15,12 +16,15 @@ from multiprocessing import Pool,cpu_count
 from itertools import repeat
 import pandas as pd 
 import numpy as np 
+import lasair
+
 
 from astropy.time import Time
 from astropy.io import fits
 from astropy.nddata import Cutout2D
 from astropy.wcs.wcs import WCS
 from astropy.utils.data import get_pkg_data_filename
+from astropy.modeling.rotations import Rotation2D
 from ztf_mag_pipeline import get_json
 
 
@@ -31,6 +35,11 @@ def convert2jd(obs_date):
         t = Time(obs_date)
         t.format = 'jd'
         return t.value
+    elif type(obs_date) is float:
+        if obs_date <= 2400000.5:
+            return obs_date + 2400000.5
+        else:
+            return obs_date
     else:
         jds = []
         for d in obs_date.tolist():
@@ -66,33 +75,45 @@ def convert_ztf(ztf_name, obj_file):
     #     collect_image(size, duration, obj_table.iloc[[i]], outdir)
 
 
-def test_valid_and_flip(path, rotation = False):
-    try:
+def test_valid(path):
+    if os.path.exists(path):
         if os.path.getsize(path) < 800:
             os.remove(path)
             return 0
         else:
-            try:
-                if rotation == False:
-                    flip_image(path)
-                else:
-                    rotate_image(path)
-            except:
-                os.remove(path)
-                return 0
             return 1
-    except:
+    else:
         return 0
+   
+    
+
+
+def test_valid_and_flip(path, rotation = False, flip = False):
+    if os.path.getsize(path) < 800:
+        os.remove(path)
+        return 0
+    else:
+        flag = 1
+        if rotation is True:
+            rotate_image(path)
+        if flip is True:
+            flag = flip_image(path)
+        if flag == 1:
+            return 1
+        else:
+            return 0
+        
+
 
 
 def path_safe(path):
-        original_umask = os.umask(0)
-        try:
-            os.makedirs(path, mode=0o775)
-        except OSError as e:
-            if e.errno != errno.EEXIST:
-                raise
-        os.umask(original_umask)
+    original_umask = os.umask(0)
+    try:
+        os.makedirs(path, mode=0o775)
+    except OSError as e:
+        if e.errno != errno.EEXIST:
+            raise
+    os.umask(original_umask)
 
 
 def create_path(path, new_folder):
@@ -133,19 +154,31 @@ def flip_image(filename):
         f = fits.open(fn, ignore_missing_end=True)[0]
         f.data = np.flip(f.data, 0)
         f.writeto(filename, overwrite=True)
+        return 1
     else:
         f = fits.open(filename,ignore_missing_end=True)
         f.verify('fix')
-        f[1].data = np.flip(f[1].data, 0)
-        f.writeto(filename, overwrite=True)
+        if f[0].data is not None:
+            f[0].data = np.flip(f[0].data, 0)
+            f.writeto(filename, overwrite=True)
+            return 1
+        elif len(f) > 1 and f[1].data is not None:
+            f[1].data = np.flip(f[1].data, 0)
+            f.writeto(filename, overwrite=True)
+            return 1
+        else:
+            os.remove(filename)
+            return 0
+
 
 
 def rotate_image(filename):
     # Used for ref image.
-    fn = get_pkg_data_filename(filename)
-    f = fits.open(fn,ignore_missing_end=True)[0]
+    # fn = get_pkg_data_filename(filename)
+    f = fits.open(filename,ignore_missing_end=True)[0]
     f.data = np.rot90(f.data, 2)
     f.writeto(filename, overwrite=True)
+  
 
 
 def check_complete(path):
@@ -158,16 +191,20 @@ def check_complete(path):
 
 
 def read_table(table, size, duration, outdir, magdir, parrallel = False):
+    ''' read a list of ZTF object'''
+
     if parrallel:
         with Pool() as pool:
-            pool.starmap(collect_image, zip(table['object_id'], table['disdate'], table['type'], repeat(size), repeat(duration), repeat(outdir), repeat(magdir)))
+            pool.starmap(collect_image_from_irsa, zip(table['object_id'], table['disdate'], table['type'], repeat(size), repeat(duration), repeat(outdir), repeat(magdir)))
     else:
         for _, row in table.iterrows():
-            collect_image(row['object_id'], row['disdate'], row['type'], size, duration, outdir, magdir)
+            collect_image_from_irsa(row['object_id'], row['disdate'], row['type'], size, duration, outdir, magdir)
 
 
 
-def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
+def collect_image_from_irsa(ztf_id, disdate, type, size, duration, outdir, magdir):
+    ''' download all existed images of a ZTF object, from irsa API.'''
+
     def add_flag(disdate, start, end):
         '''
         Add a flag to the meta:
@@ -197,7 +234,11 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
         disdate = convert2jd(disdate)
     
     if type is None:
-        label = jfile['TNS']['type']
+        if 'TNS' in jfile.keys() and 'type' in jfile['TNS']:
+            label = jfile['TNS']['type']
+            print('TNS has its classification as ', jfile['TNS']['type'])
+        else:
+            label = None
     else:
         label = type
     
@@ -325,13 +366,14 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
                             # science image:
                             
                             irsa_url = "%4s/%4s/%6s/ztf_%14s_%s_%s_c%s_%s_q%s_sciimg.fits" % (yyyy, mmdd, fracday, filefracday, field, filtercode, ccdid, imgtypecode, qid)
-                            irsa_url = '\"https://irsa.ipac.caltech.edu/ibe/data/ztf/products/sci/%s?center=%s,%s&size=%sarcmin&gzip=true\"' % (irsa_url, ra, dec, size)
-                            sci_fname = "sci_ztf_%14s_%s_%s_c%s_%s_q%s_sciimg.fits" % (filefracday, field, filtercode, ccdid, imgtypecode, qid)
+                            irsa_url = '\"https://irsa.ipac.caltech.edu/ibe/data/ztf/products/sci/%s?center=%s,%s&size=%sarcmin&gzip=false\"' % (irsa_url, ra, dec, size)
+                            sci_fname = "sci_ztf_%14s_%s_%s_c%s_%s_q%s_sciimg.fits.fz" % (filefracday, field, filtercode, ccdid, imgtypecode, qid)
                             sci_filename = obsjd_path + '/' + sci_fname
+                            print('science image url: ', irsa_url)
 
                             if not os.path.exists(sci_filename):
                                 os.system("curl -o %s %s" % (sci_filename, irsa_url))
-                                test_re = test_valid_and_flip(sci_filename)
+                                test_re = test_valid_and_flip(sci_filename, flip = True)
                                 if test_re == 0:
                                     meta_dict['f'+str(f)]['obs_with_no_sci'].append(filefracday) 
                                 else:
@@ -349,7 +391,7 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
                                     m = open(obsjd_path + "/mag_info.json",'r')
                                     mag_info = json.loads(m.read())
                                     mag_info['filefracday'] = filefracday
-                                    mag_with_img_dict['candidates_with_image']['f'+f].append(mag_info)
+                                    mag_with_img_dict['candidates_with_image']['f'+str(f)].append(mag_info)
 
                             
                             # difference image:
@@ -357,13 +399,13 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
                             diff_url = '\"https://irsa.ipac.caltech.edu/ibe/data/ztf/products/sci/%s?center=%s,%s&size=%sarcmin&gzip=true\"' % (diff_url, ra, dec, size)
                             diff_frame = "diff_ztf_%14s_%s_%s_c%s_%s_q%s_scimrefdiffimg.fits.fz" % (filefracday, field, filtercode, ccdid, imgtypecode, qid)
                             diff_filename = obsjd_path + '/' + diff_frame
+                            print('diff image url: ', diff_url)
                             
 
                             if not os.path.exists(diff_filename):
                                 os.system("curl -o %s %s" % (diff_filename, diff_url))
-                                print(diff_url)
 
-                                test_re = test_valid_and_flip(diff_filename)
+                                test_re = test_valid_and_flip(diff_filename, flip = True)
                                 if test_re == 0:
                                     meta_dict['f'+str(f)]['obs_with_no_diff'].append(filefracday)
      
@@ -375,14 +417,15 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
                         else:
                             continue
                     
-                    # reference image:
+                    # reference image: this is flipped by irsa database.
                     fieldprefix = field[:3]
-                    ref_url = '\"https://irsa.ipac.caltech.edu/ibe/data/ztf/products/ref/%s/field%s/%s/ccd%s/q%s/ztf_%s_%s_c%s_q%s_refimg.fits?center=%s,%s&size=%sarcmin&gzip=true\"' % (fieldprefix, field, filtercode, ccdid, qid, field, filtercode, ccdid, qid, ra, dec, size)
+                    ref_url = '\"https://irsa.ipac.caltech.edu/ibe/data/ztf/products/ref/%s/field%s/%s/ccd%s/q%s/ztf_%s_%s_c%s_q%s_refimg.fits?center=%s,%s&size=%sarcmin&gzip=false\"' % (fieldprefix, field, filtercode, ccdid, qid, field, filtercode, ccdid, qid, ra, dec, size)
                     ref_frame = 'ref_ztf_'+field+'_'+filtercode+'_c'+ccdid+'_q'+qid+'_refimg.fits'
                     ref_filename = path + '/' + ref_frame
+                    print('ref image url: ', ref_url)
                     if not os.path.exists(ref_filename):
                         os.system("curl -o %s %s" % (ref_filename, ref_url))
-                        print(ref_url)
+                        print(ref_filename)
                         test_re = test_valid_and_flip(ref_filename, rotation = True)
                         if test_re == 0:
                             meta_dict['f'+str(f)]['obj_with_no_ref'] = True
@@ -417,9 +460,180 @@ def collect_image(ztf_id, disdate, type, size, duration, outdir, magdir):
         return 0
     
 
+def collect_image_from_lasair(ztf_id, disdate, type, size, duration, outdir, magdir):
+    ''' download all existed images of a ZTF object, from Lasair mag JSON file.'''
+
+    def add_flag(disdate, start, end):
+        '''
+        Add a flag to the meta:
+            0: discover date within the start and end dates
+            1: discover date before the start date
+            2: discover date after the end date
+            3: no valid dates
+        '''
+        if disdate >= start and disdate < end:
+            return 0
+        elif disdate < start:
+            return 1
+        elif disdate >= end:
+            return 2
+        else:
+            return 3
+
+    print('Collecting ZTF object: ', ztf_id)
+    jfile = get_json(ztf_id, magdir)
+
+    # remove non-detection
+    temp_list = []
+    for cd in jfile['candidates']:
+        if 'candid' in cd.keys():
+            temp_list.append(cd)
+    jfile['candidates'] = temp_list
+ 
+
+    ra, dec = jfile["objectData"]["ramean"], jfile["objectData"]["decmean"]
+
+    disdate = jfile['objectData']['discMjd']
+    discFilter = jfile['objectData']['discFilter']
+
+    
+    if type is None:
+        if 'TNS' in jfile.keys() and 'type' in jfile['TNS']:
+            label = jfile['TNS']['type']
+            print('TNS has its classification as ', jfile['TNS']['type'])
+        else:
+            label = None
+    else:
+        label = type
+
+    if label is None:
+        print('Classification is no found.\n')
+
+    
+    obj_name = ztf_id
+
+    mag_cand = jfile['candidates']
+    obj_dir = outdir + '/' + obj_name 
+
+    if os.path.exists(obj_dir) and check_complete(obj_dir):
+        print('ALREADY EXIST AND COMPLETE: ', obj_name)
+        return 1
+    else:
+        obj_dir = create_path(outdir, obj_name)
+        print('Collecting images: ', obj_name)
+
+        # generate a .JSON meta file
+        meta_dict = {}
+        meta_dict['id'] = obj_name
+        meta_dict['label'] = label
+        meta_dict['ra'], meta_dict['dec'] = ra, dec 
+        meta_dict['size'] = 1
+        meta_dict['disdate'] = disdate
+
+        # generate a .JSON file for magnitudes with images provided
+        mag_with_img_dict = {}
+        mag_with_img_dict['id'] = obj_name
+        mag_with_img_dict['label'] = label
+        mag_with_img_dict['ra'] = ra
+        mag_with_img_dict['dec'] = dec
+        mag_with_img_dict['disdate'] = disdate
+        mag_with_img_dict['candidates_with_image'] = {'f1':[], 'f2':[], 'f3':[]}
+            
+
+        f_subsets=[1,2,3]
+        # f_subsets = {'g':1, 'r':2, 'i':3}
+
+        for f in f_subsets: # multithreading available
+            meta_dict['f'+str(f)] = {}
+            path = create_path(obj_dir, f)
+
+            candidates = [x for x in mag_cand if 'image_urls' in x.keys() and x['fid'] == f]
+            candidates_mjd = [x['mjd'] for x in candidates]
+
+            if len(candidates) >= 1: # if the file has observations
+                earliest = min(candidates_mjd)
+                latest = max(candidates_mjd)
+
+                start_mjd = disdate
+                end_mjd = start_mjd + duration
+
+                if start_mjd < earliest and end_mjd >= earliest:
+                    start_mjd = earliest
+                elif end_mjd < earliest:
+                    start_mjd = earliest
+                    end_mjd = earliest + duration
+                elif latest < start_mjd:
+                    start_mjd = None
+                    end_mjd = None
+                
+                meta_dict['f'+str(f)]['start'] = start_mjd
+                meta_dict['f'+str(f)]['end'] = end_mjd
+                if start_mjd is None:
+                    continue
+                
+                meta_dict['f'+str(f)]['obsnum'] = 0
+                meta_dict['f'+str(f)]['flag'] = add_flag(disdate, start_mjd, end_mjd)
+                # columns_name = df.columns
+
+                meta_dict['f'+str(f)]['multiple_ref'] = False
+                meta_dict['f'+str(f)]['obj_with_no_ref'] = False
+                meta_dict['f'+str(f)]['obs_with_no_diff'] = []
+                meta_dict['f'+str(f)]['obs_with_no_sci'] = []
+                meta_dict['f'+str(f)]['bogus'] = [] # leave the space for further operations
+                meta_dict['f'+str(f)]['withMag'] = []
+
+                count = 0
+                if len(candidates) >= 1:
+                    for c in candidates:
+                        if c['mjd'] >= start_mjd and c['mjd'] <= end_mjd: 
+                            obsjd_path = create_path(path, c["candid"])
+                            science_url  = c['image_urls']['Science']
+                            template_url = c['image_urls']['Template']
+                            difference_url = c['image_urls']['Difference']
+                            sci_name = 'sci_' + str(c['candid']) + '.fits'
+                            ref_name = 'ref_' + str(c['candid']) + '.fits'
+                            diff_name = 'diff_'+ str(c['candid']) + '.fits'
+                            sci_filename = obsjd_path + '/' + sci_name 
+                            diff_filename = obsjd_path + '/' + diff_name
+                            ref_filename = path + '/' + ref_name
+                            if not os.path.exists(sci_filename):
+                                    os.system("curl -o %s %s" % (sci_filename, science_url))
+                                    os.system("curl -o %s %s" % (ref_filename, template_url))
+                                    os.system("curl -o %s %s" % (diff_filename, difference_url))
+                            # test if images have been downloaded correctly
+                            if test_valid(sci_filename):
+                                meta_dict['f'+str(f)]['withMag'].append(c['candid'])  
+                                mag_with_img_dict['candidates_with_image']['f'+ str(f)].append(c)
+                                count += 1 
+                            else:
+                                meta_dict['f'+str(f)]['obs_with_no_sci'].append(c['candid'])
+                            if not test_valid(diff_filename):
+                                meta_dict['f'+str(f)]['obs_with_no_diff'].append(c['candid'])
+                               
+                            test_valid(ref_filename)
+                            if not os.path.exists(sci_filename) and not os.path.exists(diff_filename):
+                                os.rmdir(obsjd_path)
+
+                        else:
+                            continue
+
+                meta_dict['f'+str(f)]['obsnum'] = count
+
+        # if lasair doesn't have its images, try isra API
+        if len(meta_dict['f1']['withMag']) == 0 and len(meta_dict['f2']['withMag']) == 0:
+            print('Lasair does not have %s\' images. Trying isra API. \n'% ztf_id)
+            collect_image_from_irsa(ztf_id, disdate, type, size, duration, outdir, magdir)
+        else:
+            with open(obj_dir + "/image_meta.json", "w") as outfile:
+                json.dump(meta_dict, outfile, indent=4)
+            
+            with open(obj_dir+ '/mag_with_img.json','w') as outfile:
+                json.dump(mag_with_img_dict, outfile, indent=4)
+            print ('Done: ', obj_name)
 
 
-
+        return 0
+    
 
 
 
@@ -463,28 +677,5 @@ if __name__ == '__main__':
     
     size = 1
     duration = 100
-
-    # outdir = 'TDE_demo_set/'
-    # if os.path.isdir(outdir) == False:
-    #     os.mkdir(outdir)
-    # magdir = 'TDE_mag/'
-    # if os.path.isdir(magdir) == False:
-    #     os.mkdir(magdir)
-    
-    # obj_table = '../../data/TDE_20221017.csv'
-    # obj_table = pd.read_csv(obj_table)
-    # read_table(obj_table, size, duration, outdir, magdir, parrallel = True)
-
-    ztf_id = 'ZTF22abegjtx'
-    type = 'TDE'
-    disdate = 2458922.5
-    outdir = 'test/'
-    if os.path.isdir(outdir) == False:
-        os.mkdir(outdir)
-    magdir = 'mag_test/'
-    if os.path.isdir(magdir) == False:
-        os.mkdir(magdir)
-    collect_image(ztf_id, disdate, type, size, duration, outdir, magdir)
-
